Enable use of caching HTTP proxies
Most HTTP clients just tunnel TLS through a proxy using the HTTP CONNECT
method, which means the proxy can't provide caching services for any
HTTPS sites. Since those are increasingly common, I thought delegating
the caching functionality to a separate service wasn't a viable option.

Fortunately, httpx is one of a few client libraries that allow the
application to specify that all requests should be sent in the clear to
the proxy, which allows the proxy to provide caching or any other
services it's configured for even when accessing https:// URIs.

I've tested this with a simple Squid configuration.